(PUP-10441) Move Puppet:Util::HttpProxy to Puppet::HTTP::Proxy
Move the module to the `Puppet::HTTP` but still define the old constant to
avoid breaking modules.
