+++ lib/puppet/forge/repository.rb
@@ -1,7 +1,6 @@
 require 'puppet/ssl/openssl_loader'
 require 'digest/sha1'
 require 'uri'
-require 'puppet/util/http_proxy'
 require 'puppet/forge'
 require 'puppet/forge/errors'
 require 'puppet/network/http'
